requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
openai==1.3.0
google-generativeai==0.3.2
scikit-learn==1.3.2
numpy==1.24.3
python-dotenv==1.0.0
pydantic==2.5.0
sentence-transformers==2.2.2
langchain==0.1.0
langchain-openai==0.0.5
langchain-google-genai==0.0.6
faiss-cpu==1.7.4